🧪 Cast binomial samples to target array dtype in tests

+++ tests/test_vitalstatistics.py
@@ -34,19 +34,19 @@ class TestVitalStatistics(unittest.TestCase):
         params.V2_j_initial[:] = 0
 
         # Distribute population across various states to our liking
-        params.E_j_initial[:] = np.random.binomial(params.S_j_initial, 0.5)  # 50% of S become E
+        params.E_j_initial[:] = np.random.binomial(params.S_j_initial, 0.5).astype(params.E_j_initial.dtype)  # 50% of S become E
         params.S_j_initial -= params.E_j_initial
 
-        params.I_j_initial[:] = np.random.binomial(params.S_j_initial, 0.1)  # 10% of remaining S become I
+        params.I_j_initial[:] = np.random.binomial(params.S_j_initial, 0.1).astype(params.I_j_initial.dtype)  # 10% of remaining S become I
         params.S_j_initial -= params.I_j_initial
 
-        params.R_j_initial[:] = np.random.binomial(params.S_j_initial, 0.5)  # 50% of remaining S become R
+        params.R_j_initial[:] = np.random.binomial(params.S_j_initial, 0.5).astype(params.R_j_initial.dtype)  # 50% of remaining S become R
         params.S_j_initial -= params.R_j_initial
 
-        params.V1_j_initial[:] = np.random.binomial(params.S_j_initial, 0.1)  # 10% of remaining S become V1
+        params.V1_j_initial[:] = np.random.binomial(params.S_j_initial, 0.1).astype(params.V1_j_inital.dtype)  # 10% of remaining S become V1
         params.S_j_initial -= params.V1_j_initial
 
-        params.V2_j_initial[:] = np.random.binomial(params.S_j_initial, 0.1)  # 10% of remaining S become V2
+        params.V2_j_initial[:] = np.random.binomial(params.S_j_initial, 0.1).astype(params.V2_j_initial.dtype)  # 10% of remaining S become V2
         params.S_j_initial -= params.V2_j_initial
 
         return params
